Restructure project's directories
