+++ TagsCloud/TagsCloudVisualization/CloudLayouter/ICloudLayouter.cs
@@ -1,0 +1,9 @@
+﻿using System.Drawing;
+
+namespace TagsCloudVisualization.CloudLayouter
+{
+    public interface ICloudLayouter
+    {
+        public Rectangle PutNextRectangle(Size rectangleSize);
+    }
+}
