+++ TagsCloud/TagsCloudVisualization/ColorGenerators/GrayscaleColorGenerator.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Drawing;
-using TagsCloudVisualization.Interfaces;
 
 namespace TagsCloudVisualization.ColorGenerators
 {
