+++ TagsCloud/TagsCloudVisualization/Program.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Drawing;
 using System.Drawing.Imaging;
 using System.IO;
+using TagsCloudVisualization.CloudLayouter;
 using TagsCloudVisualization.ColorGenerators;
 
 // Disable warning https://docs.microsoft.com/ru-ru/dotnet/fundamentals/code-analysis/quality-rules/ca1416
