+++ TagsCloud/TagsCloudVisualization/TagsCloudGenerator.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
+using TagsCloudVisualization.CloudLayouter;
 
 namespace TagsCloudVisualization
 {
