+++ TagsCloud/TagsCloudVisualization/TagsCloudVisualization.csproj
@@ -6,8 +6,8 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="System.Drawing.Common" Version="6.0.0-rc.2.21480.5"/>
-        <PackageReference Include="System.Drawing.Primitives" Version="4.3.0"/>
+        <PackageReference Include="System.Drawing.Common" Version="6.0.0-rc.2.21480.5" />
+        <PackageReference Include="System.Drawing.Primitives" Version="4.3.0" />
     </ItemGroup>
 
 </Project>
