+++ TagsCloud/TagsCloudVisualization/VectorsGenerator/IVectorsGenerator.cs
@@ -1,6 +1,6 @@
 ﻿using System.Drawing;
 
-namespace TagsCloudVisualization.Interfaces
+namespace TagsCloudVisualization.VectorsGenerator
 {
     public interface IVectorsGenerator
     {
